requirements.txt
blinker==1.9.0
click==8.2.0
dataclasses==0.6
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
setuptools==78.1.1
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
zipp==3.21.0